[D7.2] Fix library dependency
* Added libtinfo.so.5 and a note to set LD_LIBRARY_PATH to work around a missing dependency. The problem originates from compiling on Ubuntu 18.04 (to support the FPGA ops) while our main execution environment is Ubuntu 20.04. Another remedy would be to install the legacy version of libtinfo:

sudo apt install libtinfo5

+++ deliverables/7.2/setup-gpu.sh
@@ -1,23 +1,26 @@
 #!/usr/bin/env bash
 
-if [[ -d bin || -d build || -d lib ]]; then
-  while true; do
-  read -p "bin, build or lib directory exist. Run build --clean first? (y/n) " yn
-  case $yn in
-    y ) echo cleaning;
-      rm -rf bin build lib
-      break;;
-    n ) echo skipping cleanup;
-      break;;
-    * ) echo invalid response;;
-  esac
-  done
-fi
+# if this is run from the source snapshot, ask to recompile
+if [[ -d src ]]; then
+  if [[ -d bin || -d build || -d lib ]]; then
+    while true; do
+    read -p "bin, build or lib directory exist. Cleanup first to rebuild? (y/n) " yn
+    case $yn in
+      y ) echo cleaning;
+        rm -rf bin build lib
+        break;;
+      n ) echo skipping cleanup;
+        break;;
+      * ) echo invalid response;;
+    esac
+    done
+  fi
 
-if ! [[ -d bin || -d build || -d lib ]]; then
-  ./build.sh --cuda
-else
-  echo skipping rebuild
+  if ! [[ -d bin || -d build || -d lib ]]; then
+    ./build.sh --cuda
+  else
+    echo skipping rebuild
+  fi
 fi
 
 ROWS=500000
